scripts: hid_configurator: Split configurator
Split configurator into functional and presentation parts.
configurator_core.py extracted from configurator.py

Jira: DESK-604, DESK-619
